+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHXEZ28ZZNV0R5KFS71Y7PPE"
+        "01KHXK2SHN23JM4MGY0ZQ0RJWF"
     )
     r"""The unique identifier of the data source"""