+++ packages/orq-rc/RELEASES.md
@@ -5869,3 +5869,13 @@ Based on:
 - [python v4.4.0-rc.19] packages/orq-rc
 ### Releases
 - [PyPI v4.4.0-rc.19] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.19 - packages/orq-rc
+
+## 2026-02-20 13:18:28
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.723.0 (2.835.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.4.0-rc.20] packages/orq-rc
+### Releases
+- [PyPI v4.4.0-rc.20] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.20 - packages/orq-rc